pyproject.toml
[build-system]
# AVOID CHANGING REQUIRES: IT WILL BE UPDATED BY PYSCAFFOLD!
requires = ["setuptools>=46.1.0", "setuptools_scm[toml]>=5"]
build-backend = "setuptools.build_meta"

[project]
name = "pisa"
version = "0.0.1"  # Placeholder for uv, actual version managed by setuptools-scm
description = "Markdownic Agent Runtime Framework"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Winsahre Tom", email = "tanwenxuan@live.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
]
dependencies = [
    "openai-agents>=0.5.0",
    "python-dotenv>=1.2.1",
    "temporalio>=1.18.2",
    "colorlog>=6.8.0",
    "arxiv>=2.1.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "crawl4ai>=0.7.6",
    "rich>=14.2.0",
    "capsolver>=1.0.7",
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "exa-py>=2.0.1",
]
[project.scripts]
pisa = "pisa.cli.main:run"

[project.optional-dependencies]
testing = [
    "setuptools",
    "pytest",
    "pytest-cov",
]
docs = [
    "sphinx>=3.2.1",
]

[project.urls]
Homepage = "https://github.com/pyscaffold/pyscaffold/"
Documentation = "https://pyscaffold.org/"

[tool.setuptools]
package-dir = {"" = "src"}
zip-safe = false

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*"]

[tool.setuptools_scm]
# For smarter version schemes and other configuration options,
# check out https://github.com/pypa/setuptools_scm
version_scheme = "no-guess-dev"

[tool.pytest.ini_options]
addopts = [
    "--cov=pisa",
    "--cov-report=term-missing",
    "--verbose",
]
norecursedirs = ["dist", "build", ".tox"]
testpaths = ["tests"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".tox",
    "build",
    "dist",
    ".eggs",
    "docs/conf.py",
]
